requirements.txt
streamlit
pandas
google-generativeai
openpyxl
pyarrow
streamlit-mic-recorder
